moved contactController into /api/v1 folder
